pyproject.toml
[project]
name = "audio-rag"
version = "0.1.0"
description = "Simple Audio RAG for beginners"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.11.0",
    "chromadb>=1.3.5",
    "compressed-tensors>=0.12.2",
    "google-genai>=1.51.0",
    "mistralai>=1.9.11",
    "openai-whisper",
    "python-dotenv>=1.0.0",
    "qwen-omni-utils>=0.0.8",
    "soundfile>=0.13.1",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchvision>=0.24.1",
    "transformers>=4.30.0",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
